+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3ClassicAirDecompileTest.java
@@ -1938,6 +1938,35 @@ public class ActionScript3ClassicAirDecompileTest extends ActionScript3Decompile
     }
 
     @Test
+    public void testSwitchBig() {
+        decompileMethod("classic_air", "testSwitchBig", "var k:int = 10;\r\n"
+                + "switch(k)\r\n"
+                + "{\r\n"
+                + "case \"A\":\r\n"
+                + "trace(\"A\");\r\n"
+                + "break;\r\n"
+                + "case \"B\":\r\n"
+                + "case \"C\":\r\n"
+                + "trace(\"BC\");\r\n"
+                + "break;\r\n"
+                + "case \"D\":\r\n"
+                + "default:\r\n"
+                + "case \"E\":\r\n"
+                + "trace(\"D-default-E\");\r\n"
+                + "break;\r\n"
+                + "case \"F\":\r\n"
+                + "trace(\"F no break\");\r\n"
+                + "case \"G\":\r\n"
+                + "trace(\"G\");\r\n"
+                + "break;\r\n"
+                + "case \"H\":\r\n"
+                + "trace(\"H last\");\r\n"
+                + "}\r\n"
+                + "trace(\"after switch\");\r\n",
+                 false);
+    }
+
+    @Test
     public void testSwitchComma() {
         decompileMethod("classic_air", "testSwitchComma", "var b:int = 5;\r\n"
                 + "var a:String = \"A\";\r\n"
@@ -1948,7 +1977,7 @@ public class ActionScript3ClassicAirDecompileTest extends ActionScript3Decompile
                 + "break;\r\n"
                 + "case \"B\":\r\n"
                 + "trace(\"is B\");\r\n"
-                + "case \"C\":\r\n"
+                + "case 7, \"C\":\r\n"
                 + "trace(\"is C\");\r\n"
                 + "}\r\n",
                  false);