+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file.
 - Highlighter nullpointer
 - AS3 search slot name Index out of bounds
 - AS text search - not being able to cancel search over multiple swf files
+- [#2486] AS3 switches detection in some cases
 
  ### Changed
 - Icon of "Deobfuscation options" menu from pile of pills to medkit